refactor(cmd_cleanup): split mod into mod+disk (#1266)
mod.rs 478 -> mod.rs 152 + disk.rs 338. Disk-related cleaners (canaries,
stale targets, target-dir cap, binary backups, corrupt DBs, snapshots,
dir_size helper) moved to disk.rs and re-exported from mod.rs.
